chore(builder): move getBrowserslistWithDefault to shared
