+++ packages/builder/builder-shared/src/getBrowserslist.ts
@@ -1,3 +1,6 @@
+import { DEFAULT_BROWSERSLIST } from './constants';
+import type { BuilderTarget, SharedBuilderConfig } from './types';
+
 // using cache to avoid multiple calls to loadConfig
 const browsersListCache = new Map<string, string[]>();
 
@@ -17,4 +20,17 @@ export async function getBrowserslist(path: string) {
   }
 
   return null;
+}
+
+export async function getBrowserslistWithDefault(
+  path: string,
+  config: SharedBuilderConfig,
+  target: BuilderTarget,
+) {
+  if (config?.output?.overrideBrowserslist) {
+    return config.output.overrideBrowserslist;
+  }
+
+  const result = await getBrowserslist(path);
+  return result || DEFAULT_BROWSERSLIST[target];
 }
